Middle of the Linked List
Link:
https://leetcode.com/problems/middle-of-the-linked-list/description/

In this solution, I have shared 2 approaches one is the "Brute Force" approach in which we first calculate the size of the Linked List by traversing through it and then finding the middle node by traversing through it again until we reach there.
Whereas, in the second approach we use are using to pointers to move quicker so once the fast pointer reaches to NULL or one step closer to NULL, the slow pointer automatically points to the middle.